+++ src/players/monte_carlo.rs
@@ -9,12 +9,12 @@ use connect_rust_graphs::graph::Graph;
 pub struct Engine {
     color: PlayerColor, 
     // Gamestates graph where gamestates have no labels
-    gamestate_graph: Graph<u32>,
+    gamestate_graph: Graph<u128>,
 
     // Contains the evaluation of each gamestate
     // The i32 being the evaluation with -1 for each loss, +1 for each win and 0 addd for each draw
     // The u32 is the total number of simulations the gamestate was involved in
-    gamestate_evaluations: HashMap<u32, (i32,u32)>,
+    gamestate_evaluations: HashMap<u128, (i32,u32)>,
 }
 
 impl Engine {
